+++ Core/Inc/robot_pow_operation.h
@@ -1,0 +1,5 @@
+#ifndef RO_P_OPE
+#define RO_P_OPE
+#include "ope_robot.h"
+#include <math.h>
+#endif
