+++ Core/Src/main.c
@@ -22,7 +22,8 @@
 
 /* Private includes ----------------------------------------------------------*/
 /* USER CODE BEGIN Includes */
-#include <math.h>
+#include "wireless_reception.h"
+#include "robot_pow_operation.h"
 /* USER CODE END Includes */
 
 /* Private typedef -----------------------------------------------------------*/
@@ -37,41 +38,18 @@
 
 /* Private macro -------------------------------------------------------------*/
 /* USER CODE BEGIN PM */
-#define RX_LENGTH 4
-#define ROBOT2_3
-// #define ROBOT2_2
-// #define ROBOT2_3
-// #define ROBOT2_4
+
 /* USER CODE END PM */
 
 /* Private variables ---------------------------------------------------------*/
-TIM_HandleTypeDef htim1;
-TIM_HandleTypeDef htim3;
-TIM_HandleTypeDef htim6;
-TIM_HandleTypeDef htim17;
 
-UART_HandleTypeDef huart1;
-UART_HandleTypeDef huart2;
 
 /* USER CODE BEGIN PV */
 uint8_t wheelConFlag;
 uint32_t timerCount;
 uint16_t stateCount;
 uint8_t timerFlag = 0;
-uint8_t U1RXbuffer;
-uint8_t controlerVarBuffer[RX_LENGTH];
-uint8_t controlerFlag;
-uint8_t con_cnt;
-typedef struct
-{
-  uint8_t buttonSW_1;
-  uint8_t buttonSW_2;
-  uint8_t buttonSW_3;
-  uint8_t buttonSW_4;
-  uint8_t toggleSW;
-  uint16_t Vartical;
-  uint16_t Horizontal;
-} inputState;
+
 /* USER CODE END PV */
 
 /* Private function prototypes -----------------------------------------------*/
@@ -110,17 +88,7 @@ void HAL_UART_RxCpltCallback(UART_HandleTypeDef *huart)
   /*受信した値を格納*/
   if (huart == &huart1)
   {
-    controlerVarBuffer[con_cnt] = U1RXbuffer;
-    if (controlerVarBuffer[0] == 0x80)
-    {
-      con_cnt++;
-    }
-    if (con_cnt == RX_LENGTH)
-    {
-      con_cnt = 0;
-      controlerFlag = 1;
-      HAL_GPIO_WritePin(LED_GPIO_Port, LED_Pin, GPIO_PIN_SET);
-    }
+    uart_reception();
   }
 }
 /* USER CODE END 0 */
@@ -575,252 +543,6 @@ static void MX_GPIO_Init(void)
 
 /* USER CODE BEGIN 4 */
 
-/**
- * @brief Control functions for differential two-wheel
- * @retval None
- *
- * @param Horizontal Horizontal axis value of stick
- * @param vertical Vertical axis value of stick
- */
-void WheelPowControl(double Horizontal, double Vartical)
-{
-  const int STICK_CENTER_POSITION = 0x40;
-  double leftWheelPow;
-  double rightWheelPow;
-  double radian;
-  Horizontal -= (double)STICK_CENTER_POSITION;
-  Vartical -= (double)STICK_CENTER_POSITION;
-  radian = atan2(Horizontal, Vartical);
-  double powerGain = (hypot(Vartical, Horizontal) / (2 * STICK_CENTER_POSITION));
-  powerGain = (powerGain >= 1) ? 1 : powerGain;
-  rightWheelPow = 500 + ((powerGain * 500) * sin(radian - M_3PI_4));
-  leftWheelPow = 500 + ((powerGain * 500) * sin(radian + M_3PI_4));
-  __HAL_TIM_SET_COMPARE(&htim3, TIM_CHANNEL_1, rightWheelPow);
-  __HAL_TIM_SET_COMPARE(&htim3, TIM_CHANNEL_2, leftWheelPow);
-}
-
-void DecodeControlerVarBuffer(uint8_t *controlerVarBuffer, inputState *Data)
-{
-  cntState.Horizontal = controlerVarBuffer[1];
-  cntState.Vartical = controlerVarBuffer[2];
-  if (controlerVarBuffer[3] & (1 << 0))
-  {
-    Data->buttonSW_1 = 1;
-  }
-  else
-  {
-    Data->buttonSW_1 = 0;
-  }
-  if (controlerVarBuffer[3] & (1 << 1))
-  {
-    Data->buttonSW_2 = 1;
-  }
-  else
-  {
-    Data->buttonSW_2 = 0;
-  }
-  if (controlerVarBuffer[3] & (1 << 2))
-  {
-    Data->buttonSW_3 = 1;
-  }
-  else
-  {
-    Data->buttonSW_3 = 0;
-  }
-  if (controlerVarBuffer[3] & (1 << 3))
-  {
-    Data->buttonSW_4 = 1;
-  }
-  else
-  {
-    Data->buttonSW_4 = 0;
-  }
-  if (controlerVarBuffer[3] & (1 << 4))
-  {
-    Data->toggleSW = 1;
-  }
-  else
-  {
-    Data->toggleSW = 0;
-  }
-}
-
-#ifdef ROBOT2_1
-/**
- * @brief This fanction is proglam for the robot2-X
- * @note for the robot2-1 (collecting the box)
- * @retval None
- *
- * @param Data
- */
-void IndividualOpelation(inputState *Data)
-{
-  static uint16_t powerA = 500; // for locking mechanism
-  static uint16_t powerB = 500; // for collection arm
-  static uint16_t powerC = 500; // for vacuume pump
-  int limSwState1 = HAL_GPIO_ReadPin(LimitSW1_GPIO_Port, LimitSW1_Pin);
-  int limSwState2 = HAL_GPIO_ReadPin(LimitSW2_GPIO_Port, LimitSW2_Pin);
-  uint32_t encoderVal = TIM1->CNT; //(0~2000)
-  static uint8_t swState1, swState2, lastSwState1, lastSwState2;
-
-  // HAL_UART_Transmit(&huart2, &encoderVal, sizeof(encoderVal), 0xFFFF);
-
-  /* for locking mechanism */
-  if (Data->buttonSW_4 != lastSwState1)
-  {
-    if (Data->buttonSW_4)
-    {
-      swState1 = !swState1;
-    }
-    lastSwState1 = Data->buttonSW_4;
-  }
-  if (swState1 && !limSwState1)
-  {
-    powerA = 0;
-  }
-  else if (!swState1 && !limSwState2)
-  {
-    powerA = 1000;
-  }
-  else
-  {
-    powerA = 500;
-  }
-
-  /* for collection arm */
-  if (Data->buttonSW_1 != Data->buttonSW_2)
-  {
-    if (Data->buttonSW_1 && encoderVal <= 1000 || encoderVal >= 1900)
-    {
-      powerB = 800;
-    }
-    else if (Data->buttonSW_2 && encoderVal >= 50)
-    {
-      powerB = 200;
-    }
-    else
-    {
-      powerB = 500;
-    }
-  }
-  else
-  {
-    powerB = 500;
-  }
-
-  /* for vacuume pump */
-  if (Data->buttonSW_3 != lastSwState2)
-  {
-    if (Data->buttonSW_3)
-    {
-      swState2 = !swState2;
-    }
-    lastSwState2 = Data->buttonSW_3;
-  }
-  if (swState2)
-  {
-    powerC = 1000;
-  }
-  else
-  {
-    powerC = 500;
-  }
-
-  /* set duty ratio */
-  __HAL_TIM_SET_COMPARE(&htim3, TIM_CHANNEL_3, powerA);
-  __HAL_TIM_SET_COMPARE(&htim3, TIM_CHANNEL_4, powerB);
-  __HAL_TIM_SET_COMPARE(&htim17, TIM_CHANNEL_1, powerC);
-}
-#endif /*ROBOT2_1*/
-
-#ifdef ROBOT2_2
-/**
- * @brief This fanction is proglam for the robot2-X
- * @note for the robot2-2 (collecting the ball)
- * @retval none
- *
- * @param Data
- */
-void IndividualOpelation(inputState *Data)
-{
-  static uint16_t powerA = 500; // for belt
-  static uint16_t powerB = 500; // for roller
-
-  if (Data->buttonSW_1 != Data->buttonSW_2)
-  {
-    if (Data->buttonSW_1)
-    {
-      powerA = 400;
-      powerB = 350;
-    }
-    else if (Data->buttonSW_2)
-    {
-      powerA = 600;
-      powerB = 650;
-    }
-    else if (Data->buttonSW_3)
-    {
-      powerA = 0;
-      powerB = 350;
-    }
-  }
-  else
-  {
-    powerA = 500;
-    powerB = 500;
-  }
-
-  __HAL_TIM_SET_COMPARE(&htim3, TIM_CHANNEL_3, powerA);
-  __HAL_TIM_SET_COMPARE(&htim3, TIM_CHANNEL_4, powerB);
-}
-#endif /*ROBOT2_2*/
-
-#ifdef ROBOT2_3
-/**
- * @brief This fanction is proglam for the robot2-X
- * @note for the robot2-3 (landing only)
- * @retval None
- * @param Data
- */
-void IndividualOpelation(inputState *Data)
-{
-  static uint16_t powerA = 500;
-  if (Data->buttonSW_1 != Data->buttonSW_2)
-  {
-    if (Data->buttonSW_1)
-    {
-      powerA = 0;
-    }
-    if (Data->buttonSW_2)
-    {
-      powerA = 1000;
-    }
-  }
-  else
-  {
-    powerA = 500;
-  }
-
-  __HAL_TIM_SET_COMPARE(&htim3, TIM_CHANNEL_3, powerA);
-}
-#endif /*ROBOT2_3*/
-
-#ifdef ROBOT2_4
-/**
- * @brief This fanction is proglam for the robot2-X
- * @note for the robot2-4 (not use this function)
- * @retval None
- * @param paramA
- * @param paramB
- */
-void IndividualOpelation(inputState *Data)
-{
-  static uint16_t powerA = 500;
-  static uint16_t powerB = 500;
-  __HAL_TIM_SET_COMPARE(&htim3, TIM_CHANNEL_3, powerA);
-  __HAL_TIM_SET_COMPARE(&htim3, TIM_CHANNEL_4, powerB);
-}
-#endif /*ROBOT2_4*/
 /* USER CODE END 4 */
 
 /**
